Remove commented code
