+++ json-files/author.json
@@ -1,1 +1,1 @@
-[{"first_name":"Fred","last_name":"Don"}]
+[{"first_name":"Fred","last_name":"Don"},{"first_name":"Alex","last_name":"John"}]
