+++ json-files/game.json
@@ -1,1 +1,1 @@
-[{"publish_date":null,"multiplayer":true,"last_played_at":"2022-01-01","archived":null,"first_name":"Fred"}]
+[{"publish_date":null,"multiplayer":true,"last_played_at":"2022-01-01","archived":null,"first_name":"Fred"},{"publish_date":"2020-01-01","multiplayer":true,"last_played_at":"2022-02-02","archived":true,"first_name":"Alex"}]
