+++ json-files/genre.json
@@ -1,1 +1,1 @@
-[{"name":"Afrobeats"},{"name":"RnB"}]
+[{"name":"Afrobeats"},{"name":"RnB"},{"name":"RnB"}]
