+++ json-files/music_album.json
@@ -1,1 +1,1 @@
-[{"name":"african Giant","publish_date":"2019-10-20","on_spotify":true},{"name":"Tears and Goosebumps","publish_date":"2020-06-02","on_spotify":true}]
+[{"name":"african Giant","publish_date":"2019-10-20","on_spotify":true},{"name":"Tears and Goosebumps","publish_date":"2020-06-02","on_spotify":true},{"name":"MIL","publish_date":"2021-07-08","on_spotify":true}]
