+++ storage.rb
@@ -1,21 +1,6 @@
 require 'json'
 
 module Storage
-  # def load_music_album_to_file
-  #   items = MusicAlbum.new(name, publish_date, on_spotify)
-  #   items.each do |item|
-  #     @musicalbums << item
-  #   end
-  # end
-
-  # def load_genre_to_file
-  #   items = Genre.new(name)
-  #   items.each do |item|
-  #     @genre << item
-  #   end
-  # end
-
-  # saving to the json files
   def save_state
     puts @musicalbums
     items = @musicalbums.map do |music|
@@ -35,7 +20,6 @@ module Storage
   def read_music_album
     file = File.read('./json-files/music_album.json')
     read_music = JSON.parse(file)
-    # puts read_music
 
     read_music.each do |music|
       @musicalbums << MusicAlbum.new(music['name'], music['publish_date'], music['on_spotify'])
@@ -47,7 +31,6 @@ module Storage
   def read_genre
     file = File.read('./json-files/genre.json')
     read_each_genre = JSON.parse(file)
-    # puts read_each_genre
 
     read_each_genre.each do |genre|
       @genre << Genre.new(genre['name'])
